Refactor stems.py for more readable code
The changes involve substituting hard-coded values with CONSTANTS for straight and curved glyph names which enhance readability. The `get_stems()` function was also renamed to `recalc_stems()`, providing a more accurate and understanding name. Furthermore, the values of glyphs were changed to the variables STRAIGHT and CURVED in the related function calls for more consistent code. These changes further contribute to better maintainability of our codebase.