Add GlyphBounds TypedDict (#63)
* Add GlyphBounds TypedDict and reformat get_glyph_bounds methods

This commit introduces a new TypedDict, GlyphBounds, to encapsulate the information on glyph bounds in a more structured way. With this new class, the `get_glyph_bounds` and `get_glyphs_bounds` methods are now returning GlyphBounds instances. This provides a clearer, more readable way to describe the bounds of a glyph and affords better type checking, improving the reliability of this piece of the code.