Refactor variable names for better readability
Variable names within the methods related to calculation of glyph zones were updated for better readability and understanding. The 'zone_data' was renamed to 'data' for brevity. The 'lower_case_letters' var was changed to 'lowercase_letters', similarly for 'lower_case_ascenders' and 'lower_case_descenders' for better name consistency. The 'descender_glyphs' now exclude ‘f’ instead of ‘j’ for accurate glyph classification. This renaming helps in understanding the code and variables better.