Optimize glyph zone calculations
This revision improves the code for calculating Glyph zones in the ps_recalc_zones.py file. List comprehension was replaced with Python’s set operations for better efficiency and readability when finding the list of glyphs to include in zone calculations. 'f' and 'j' characters were factored out of the descender and x-height glyphs, and 't' character was excluded from ascender glyphs, further improving accuracy.