Add functionality to convert OpenType font to TrueType in font.py
A new method has been introduced to the Font class in font.py to convert OpenType fonts to TrueType. An associated new file, foundrytools_cli_2/lib/ttf/from_otf.py has also been added with helper functions and typing details. Furthermore, the function raises a ValueError if the font is already in TrueType and a NotImplementedError if the font is variable.