Update optional parameters in font hinting methods
The 'reference_font' parameter in various font hinting methods found in the 'autohint.py' and 'afdko_tools.py' files has been updated to accept 'None' as a value. The static typing was also updated in these files from direct import to a more generalized 'typing as t', improving overall flexibility and readability of the code.