Dev (#99)
* Move contour correction step post font saving

The contour correction step has been moved to occur after the font is saved to the output file. This change ensures that the contour correction affects the 'otf' object and not the initial 'font' object, improving the accuracy and efficiency of the code.

* Update method to delete Mac-specific names

Changed method 'del_names' to accept an iterable instead of a tuple and 'del_mac_names' to delete all Mac-specific names by default. The update allows flexibility in deleting specific names and more control over which names to keep, thereby enhancing Pyfontaine's capabilities to handle Macintosh-specific font names.