Simplify otf autohinting process and function arguments
The code has been refactored to streamline the process of applying hints to an OpenType-PS font file and to simplify the arguments passed to the relevant functions. This is achieved by creating an ACOptions object to hold all options used in hinting. Code checks, validations, and temporary file handling have been adjusted accordingly.