Update lint settings and optimize sets usage
Lint settings in the .ruff.toml file were updated, including the addition of 'C' to the selected categories. Also, optimized the usage of set creation in lib/tables/name.py and snippets/name.py by using set comprehension directly instead of using a for loop. The changes aim to ensure proper linting preference and to improve code readability and performance.