Set modified flag after autohinting font
Font modification flag is now set to true after the autohinting process in the autohint.py file. This change has been added to ensure the program records and acknowledges any modifications made during the autohinting process. It improves the accuracy of status tracking and integrity of the font modification process.