Refactor attribute assignment in font.py
In the font.py script, the method of assigning attributes for the different variations of font properties has been refined. Instead of manually accessing and setting each attribute, the code now uses the setattr() built-in function to achieve this in a cleaner and more efficient way. This change affects how the properties related to "OtherBlues", "BlueValues", "StdHW" and "StdVW" are set.