Change font-to-SFNT call to resetting font flavor in 'get_file_to_process'
The previous implementation used 'font.to_sfnt' method to convert the font to SFNT format if certain conditions were met, which was not the intended functionality. This commit changes it to 'font.ttfont.flavor = None', which is more correct because the 'font.modified' attribute is not changed (which happens with 'font.to_sfnt').