Modify return type in recalc_max_context function
The return type for the `recalc_max_context` function in the `OS/2` table has been changed from int to None. This function now directly modifies the `max_context` value of the table, instead of returning the recalculated value. This change simplifies the snippet that uses this function by removing the need to manually assign the returned value to `max_context`.