Refactor 'os_2.py' to use more precise conditions for setting bits
This commit refactors the condition used when setting bits 7, 8, and 9 of the 'OS/2.fsSelection' field in 'os_2.py'. The new condition checks not only if the OS/2 table version is less than 4, but also if the new value to be set is True, thereby preventing unnecessary errors. The logger module was also imported to this file.